[IMP] test-db scrpts

+++ install-scripts/openerp-install/test-db/openerp-test-db.sh
@@ -34,14 +34,22 @@ checkRoot
 
 # Stop copy server
 /etc/init.d/openerp-server-$3 stop
-killall -s KILL openerp_$3
+server_name=openerp_$3
+proc_name=${server_name:0:15}
+if [[ `ps -A | grep -c $proc_name` -gt 0 ]]; then
+	killall -s KILL openerp_$3
+fi
+
 # Re-create the database
 sudo -u postgres psql -c "DROP DATABASE $2"
 sudo -u postgres psql -c "CREATE DATABASE $2"
+
 # Close connections do copy db
 sudo -u postgres psql -c "SELECT procpid FROM pg_stat_activity where datname='$2'" | grep -e "[0-9]$" | while read line; do kill $line; kill -s KILL $line; done
+
 # Copy database
 sudo -u postgres pg_dump --format=c --no-owner $1 | sudo -u postgres pg_restore --no-owner --dbname=$2
+
 # Restart copy server
 /etc/init.d/openerp-server-$3 start
 sudo -u postgres psql -c "ALTER DATABASE $2 OWNER TO openerp_$3"
